WEB-4460 abstract out PaginationController
